キャッシュのクラスを別ファイルで定義
